+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## Unreleased
 
+## [v0.10.0](https://github.com/thibaudcolas/curlylint/releases/tag/v0.10.0) 2020-05-21
+
 ### Added
 
 - Add `--print-config` CLI flag to print the configuration for the given file.